로그인 기능 구현 연습

+++ CAMO_Spring/src/main/java/jcw/CAMO_Spring/domain/Review.java
@@ -11,7 +11,7 @@ public class Review {
     private Long reviewId; //리뷰번호(PK)
     private Long memberId; //회원번호(FK)
     private Long cafeId; //카페번호(FK)
-    private int starRating; //별점
+    private double starRating; //별점
     private String description; //리뷰 본문
     private Date date; //작성날짜
 }
